Refactored the SQL datasources to use the template.go instead
