+++ storage/sqlStorages/mysql/datasource.go
@@ -1,118 +1,74 @@
 package mysql
 
 import (
-	"context"
 	"database/sql"
 	"fmt"
 
-	sqlStorages "github.com/cube2222/octosql/storage/sqlStorages"
+	"github.com/cube2222/octosql"
+	"github.com/cube2222/octosql/storage/sqlStorages"
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/pkg/errors"
 
-	"github.com/cube2222/octosql"
 	"github.com/cube2222/octosql/config"
-	"github.com/cube2222/octosql/execution"
 	"github.com/cube2222/octosql/physical"
-	"github.com/cube2222/octosql/physical/metadata"
 )
 
-var availableFilters = map[physical.FieldType]map[physical.Relation]struct{}{
-	physical.Primary: {
-		physical.Equal:        {},
-		physical.NotEqual:     {},
-		physical.MoreThan:     {},
-		physical.LessThan:     {},
-		physical.GreaterEqual: {},
-		physical.LessEqual:    {},
-		physical.Like:         {},
-	},
-	physical.Secondary: {
-		physical.Equal:        {},
-		physical.NotEqual:     {},
-		physical.MoreThan:     {},
-		physical.LessThan:     {},
-		physical.Like:         {},
-		physical.GreaterEqual: {},
-		physical.LessEqual:    {},
-	},
+type MySQLTemplate struct{}
+
+func (t MySQLTemplate) GetAvailableFilters() map[physical.FieldType]map[physical.Relation]struct{} {
+	return map[physical.FieldType]map[physical.Relation]struct{}{
+		physical.Primary: {
+			physical.Equal:        {},
+			physical.NotEqual:     {},
+			physical.MoreThan:     {},
+			physical.LessThan:     {},
+			physical.GreaterEqual: {},
+			physical.LessEqual:    {},
+			physical.Like:         {},
+		},
+		physical.Secondary: {
+			physical.Equal:        {},
+			physical.NotEqual:     {},
+			physical.MoreThan:     {},
+			physical.LessThan:     {},
+			physical.Like:         {},
+			physical.GreaterEqual: {},
+			physical.LessEqual:    {},
+		},
+	}
 }
 
-type DataSource struct {
-	db           *sql.DB
-	stmt         *sql.Stmt
-	placeholders []execution.Expression
-	alias        string
+func (t MySQLTemplate) GetIPAddress(dbConfig map[string]interface{}) (string, int, error) {
+	return config.GetIPAddress(dbConfig, "address", config.WithDefault([]interface{}{"localhost", 3306}))
 }
+
+func (t MySQLTemplate) GetConnection(user, password, host, dbName string, port int) (*sql.DB, error) {
+	// Build dsn
+	mysqlInfo := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?parseTime=true", user, password, host, port, dbName)
+
+	db, err := sql.Open("mysql", mysqlInfo)
+	if err != nil {
+		return nil, errors.Wrap(err, "couldn'template open connection to postgres database")
+	}
+
+	return db, nil
+}
+
+func (t MySQLTemplate) GetPlaceholders(alias string) sqlStorages.PlaceholderMap {
+	return newMySQLPlaceholders(alias)
+}
+
+var template MySQLTemplate = struct{}{}
 
 func NewDataSourceBuilderFactory(primaryKeys []octosql.VariableName) physical.DataSourceBuilderFactory {
-	return physical.NewDataSourceBuilderFactory(
-		func(ctx context.Context, matCtx *physical.MaterializationContext, dbConfig map[string]interface{}, filter physical.Formula, alias string) (execution.Node, error) {
-			// Get execution configuration
-			host, port, err := config.GetIPAddress(dbConfig, "address", config.WithDefault([]interface{}{"localhost", 3306}))
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't get address")
-			}
-			user, err := config.GetString(dbConfig, "user")
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't get user")
-			}
-			databaseName, err := config.GetString(dbConfig, "databaseName")
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't get databaseName")
-			}
-			tableName, err := config.GetString(dbConfig, "tableName")
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't get tableName")
-			}
-			password, err := config.GetString(dbConfig, "password")
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't get password")
-			}
-
-			// Build dsn
-			mysqlInfo := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?parseTime=true", user, password, host, port, databaseName)
-
-			db, err := sql.Open("mysql", mysqlInfo)
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't open connection to postgres database")
-			}
-
-			placeholders := newMySQLPlaceholders(alias)
-
-			//create a query with placeholders to prepare a statement from a physical formula
-			query := sqlStorages.FormulaToSQL(filter, placeholders)
-			query = fmt.Sprintf("SELECT * FROM %s %s WHERE %s", tableName, alias, query)
-
-			stmt, err := db.Prepare(query)
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't prepare db for query")
-			}
-
-			//materialize the created mySQLPlaceholders
-			execAliases, err := placeholders.MaterializePlaceholders(matCtx)
-
-			if err != nil {
-				return nil, errors.Wrap(err, "couldn't materialize mySQLPlaceholders")
-			}
-
-			return &DataSource{
-				stmt:         stmt,
-				placeholders: execAliases,
-				alias:        alias,
-				db:           db,
-			}, nil
-		},
-		primaryKeys,
-		availableFilters,
-		metadata.BoundedDoesntFitInLocalStorage,
-	)
+	return sqlStorages.NewDataSourceBuilderFactoryFromTemplate(template)(primaryKeys)
 }
 
 // NewDataSourceBuilderFactoryFromConfig creates a data source builder factory using the configuration.
 func NewDataSourceBuilderFactoryFromConfig(dbConfig map[string]interface{}) (physical.DataSourceBuilderFactory, error) {
 	primaryKeysStrings, err := config.GetStringList(dbConfig, "primaryKeys", config.WithDefault([]string{}))
 	if err != nil {
-		return nil, errors.Wrap(err, "couldn't get primaryKeys")
+		return nil, errors.Wrap(err, "couldn'template get primaryKeys")
 	}
 	var primaryKeys []octosql.VariableName
 	for _, str := range primaryKeysStrings {
@@ -120,86 +76,4 @@ func NewDataSourceBuilderFactoryFromConfig(dbConfig map[string]interface{}) (phy
 	}
 
 	return NewDataSourceBuilderFactory(primaryKeys), nil
-}
-
-func (ds *DataSource) Get(ctx context.Context, variables octosql.Variables) (execution.RecordStream, error) {
-	values := make([]interface{}, 0)
-
-	for i := range ds.placeholders {
-		expression := ds.placeholders[i]
-
-		//since we have an execution expression, then we can evaluate it given the variables
-		value, err := expression.ExpressionValue(ctx, variables)
-		if err != nil {
-			return nil, errors.Wrap(err, "couldn't get actual value from variables")
-		}
-
-		values = append(values, value.ToRawValue())
-	}
-
-	rows, err := ds.stmt.QueryContext(ctx, values...)
-	if err != nil {
-		return nil, errors.Wrap(err, "couldn't query statement")
-	}
-
-	columns, err := rows.Columns()
-	if err != nil {
-		return nil, errors.Wrap(err, "couldn't get columns from rows")
-	}
-
-	return &RecordStream{
-		rows:    rows,
-		columns: columns,
-		isDone:  false,
-		alias:   ds.alias,
-	}, nil
-
-}
-
-type RecordStream struct {
-	rows    *sql.Rows
-	columns []string
-	isDone  bool
-	alias   string
-}
-
-func (rs *RecordStream) Close() error {
-	err := rs.rows.Close()
-	if err != nil {
-		return errors.Wrap(err, "Couldn't close underlying SQL rows")
-	}
-
-	return nil
-}
-
-func (rs *RecordStream) Next(ctx context.Context) (*execution.Record, error) {
-	if rs.isDone {
-		return nil, execution.ErrEndOfStream
-	}
-
-	if !rs.rows.Next() {
-		rs.isDone = true
-		return nil, execution.ErrEndOfStream
-	}
-
-	cols := make([]interface{}, len(rs.columns))
-	colPointers := make([]interface{}, len(cols))
-	for i := range cols {
-		colPointers[i] = &cols[i]
-	}
-
-	if err := rs.rows.Scan(colPointers...); err != nil {
-		return nil, errors.Wrap(err, "couldn't scan row")
-	}
-
-	resultMap := make(map[octosql.VariableName]octosql.Value)
-
-	fields := make([]octosql.VariableName, len(rs.columns))
-	for i, columnName := range rs.columns {
-		newName := octosql.NewVariableName(fmt.Sprintf("%s.%s", rs.alias, columnName))
-		fields[i] = newName
-		resultMap[newName] = octosql.NormalizeType(cols[i])
-	}
-
-	return execution.NewRecord(fields, resultMap), nil
 }
